Initial commit
This commit adds a simple version of tinycat with the bare
minimum functionality.

Tinycat is a lightweight (<100K) netcat alternative which is statically
linked (against musl).
Tinycat exclusively deals with TCP, and can only handle IP and UNIX
domain sockets.

The netcat options -k, -l and -U are implemented and they have the same
meaning (refer to the help message for details).